update deps, fix issue where going to the info tab of the app settings would lock the user in place
