+++ package.json
@@ -40,13 +40,13 @@
     "@react-native-community/netinfo": "^11.4.1",
     "@react-native-masked-view/masked-view": "^0.3.2",
     "@react-native-vector-icons/material-design-icons": "^12.3.0",
-    "@react-navigation/bottom-tabs": "^7.4.9",
-    "@react-navigation/material-top-tabs": "^7.3.8",
-    "@react-navigation/native": "^7.1.18",
-    "@react-navigation/native-stack": "^7.3.28",
+    "@react-navigation/bottom-tabs": "7.5.0",
+    "@react-navigation/material-top-tabs": "7.3.9",
+    "@react-navigation/native": "7.1.18",
+    "@react-navigation/native-stack": "7.5.0",
     "@sentry/react-native": "7.1.0",
     "@shopify/flash-list": "^2.1.0",
-    "@tamagui/config": "1.135.0",
+    "@tamagui/config": "1.135.4",
     "@tanstack/query-async-storage-persister": "5.89.0",
     "@tanstack/react-query": "5.89.0",
     "@tanstack/react-query-persist-client": "5.89.0",
@@ -76,15 +76,15 @@
     "react-native-haptic-feedback": "^2.3.3",
     "react-native-linear-gradient": "^2.8.3",
     "react-native-mmkv": "3.3.3",
-    "react-native-nitro-image": "^0.6.1",
-    "react-native-nitro-modules": "^0.30.2",
+    "react-native-nitro-image": "^0.8.1",
+    "react-native-nitro-modules": "^0.31.1",
     "react-native-nitro-ota": "^0.2.3",
     "react-native-nitro-web-image": "^0.6.1",
     "react-native-ota-hot-update": "2.3.1",
     "react-native-pager-view": "^6.9.1",
     "react-native-reanimated": "4.1.3",
     "react-native-safe-area-context": "^5.6.1",
-    "react-native-screens": "4.17.1",
+    "react-native-screens": "4.18.0",
     "react-native-swipeable-item": "^2.0.9",
     "react-native-text-ticker": "^1.15.0",
     "react-native-toast-message": "^2.3.3",
@@ -94,7 +94,7 @@
     "react-native-worklets": "0.6.1",
     "ruby": "^0.6.1",
     "scheduler": "^0.26.0",
-    "tamagui": "1.135.0",
+    "tamagui": "1.135.4",
     "zustand": "^5.0.8"
   },
   "devDependencies": {
