--- src/components/Settings/components/info/storage.tsx
@@ -1,9 +1,0 @@
-import { View } from 'react-native'
-import { Text } from '../../../Global/helpers/text'
-export default function InfoTabStorage() {
-	return (
-		<View>
-			<Text>Storage</Text>
-		</View>
-	)
-}
