--- src/components/Settings/components/info/types.d.ts
@@ -1,7 +1,0 @@
-export type InfoTabStackParamList = {
-	Index: undefined
-	Storage: undefined
-}
-
-export type InfoTabNativeStackNavigationProp =
-	NativeNativeStackNavigationProp<InfoTabStackParamList>
